Remove bias estimation from bootstrap module (#176)
`bootstrap.bias` and `bootstrap.bias_corrected` are deprecated.

The bias estimation with the bootstrap seem to work in our tests, but it
is not something that people in the field do, and there is no proof that
this is the proper way.

To be on the safe side, we should not offer this. People should use the
jacknife for bias correction.