Remove resample.bootstrap.bias and resample.bootstrap.bias_corrected (#181)
The implementation of resample.bootstrap.bias was removed because of
several flaws.

It was found to give very wrong results when used with
`method="extended"` for parameters sensitive to the sample size. It was
only an approximation to the correct implementation of the bias formula,
valid only in the limit of very large bootstrap samples. This was not
properly stated in the documentation. It is not obvious how this general
flaw can be fixed in a reasonable manner, so the decision is to remove
the bias estimation via bootstrap altogether.